// blaze-integrator-app/projects/myapp/src/presentation/app/pages/add-component/add-component-enums.ts
export enum componentKeys {
  sftp = 'sftp',
  cases = 'cases',
  group = 'group',
  switch = 'switch',
  tree = 'tree',
  treeCondition = 'treeCondition',
  single = 'single',
  output = 'output',
  outputs = 'outputs',
  sequential = 'sequential',
  fallback = 'fallback',
  processors = 'processors',
  condition = 'condition',
}